ai-service/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
pillow
python-multipart
python-dotenv
huggingface_hub
